Move memory viz templates into separate javascript files
[ghstack-poisoned]
