+++ torch/cuda/_memory_viz.py
@@ -586,355 +586,9 @@ _memory_over_time_template = r"""
 <head></head>
 <body>
 <script type="module">
-import * as d3 from "https://cdn.jsdelivr.net/npm/d3@7.7.0/+esm";
-import {schemeTableau10} from "https://cdn.skypack.dev/d3-scale-chromatic@3";
-import {axisLeft} from "https://cdn.skypack.dev/d3-axis@3";
-import {scaleLinear} from "https://cdn.skypack.dev/d3-scale@4";
-import {zoom, zoomIdentity} from "https://cdn.skypack.dev/d3-zoom@3";
-import {brushX} from "https://cdn.skypack.dev/d3-brush@3";
-
+import {main} from "./MemoryPlot.js"
 let alloc_data = $PLOT_DATA
-
-function process_alloc_data(max_entries) {
-    let current = []
-    let current_data = []
-    let data = []
-    let max_size = 0
-
-    let total_mem = 0
-    let total_summarized_mem = 0
-    let timestep = 0
-
-    let max_at_time = []
-
-
-    let summarized_mem = {elem: 'summarized', timesteps: [], offsets: [total_mem], size: [], color: 0}
-    let summarized_elems = {}
-
-    function advance(n) {
-        summarized_mem.timesteps.push(timestep)
-        summarized_mem.offsets.push(total_mem)
-        summarized_mem.size.push(total_summarized_mem)
-        timestep += n
-        for (let i = 0; i < n; i++) {
-            max_at_time.push(total_mem + total_summarized_mem)
-        }
-    }
-
-    let sizes = alloc_data.elements_size.map((x, i) => [x, i]).sort(([x, xi], [y, yi]) => y - x)
-
-    let draw_elem = {}
-    for (const [s, e] of sizes.slice(0, max_entries)) {
-        draw_elem[e] = true
-    }
-
-    function add_allocation(elem) {
-        let size = alloc_data.elements_size[elem]
-        current.push(elem)
-        let color = elem
-        if (alloc_data.elements_category !== null) {
-            color = alloc_data.elements_category[elem]
-        }
-        let e = {elem: elem, timesteps: [timestep], offsets: [total_mem], size: size, color: color}
-        current_data.push(e)
-        data.push(e)
-        total_mem += size
-    }
-
-    for (const elem of alloc_data.initially_allocated) {
-        if (elem in draw_elem) {
-            add_allocation(elem)
-        } else {
-            total_summarized_mem += alloc_data.elements_size[elem]
-            summarized_elems[elem] = true
-        }
-    }
-
-    for (const action of alloc_data.actions) {
-        const elem = action
-        const size = alloc_data.elements_size[elem]
-        if ( !(elem in draw_elem)) {
-            if (elem in summarized_elems) {
-                advance(1)
-                total_summarized_mem -= size
-                summarized_elems[elem] = null
-            } else {
-                total_summarized_mem += size
-                summarized_elems[elem] = true
-                advance(1)
-            }
-            continue
-        }
-        const idx = current.findLastIndex(x => x === elem)
-        // first time we see an action we add it
-        // second time we remove it
-        if (idx == -1) {
-            add_allocation(elem)
-            advance(1)
-        } else {
-            advance(1)
-            const removed = current_data[idx]
-            removed.timesteps.push(timestep)
-            removed.offsets.push(removed.offsets.at(-1))
-            current.splice(idx, 1)
-            current_data.splice(idx, 1)
-
-            if (idx < current.length) {
-                for (let j = idx; j < current.length; j++) {
-                    const e = current_data[j]
-                    e.timesteps.push(timestep)
-                    e.offsets.push(e.offsets.at(-1))
-                    e.timesteps.push(timestep + 3)
-                    e.offsets.push(e.offsets.at(-1) - size)
-                }
-                advance(3)
-            }
-            total_mem -= size
-        }
-        max_size = Math.max(total_mem + total_summarized_mem, max_size)
-    }
-
-    for (const elem of current_data) {
-        elem.timesteps.push(timestep)
-        elem.offsets.push(elem.offsets.at(-1))
-    }
-    data.push(summarized_mem)
-
-    return {
-        max_size: max_size,
-        allocations_over_time: data,
-        max_at_time: max_at_time,
-        summarized_mem: summarized_mem,
-        context_for_id:  (elem) => {
-            let strings = []
-            let id = alloc_data.elements_info[elem]
-            while (id !== null) {
-                const [sid, next_id] = alloc_data.suffix_table[id]
-                strings.push(alloc_data.string_table[sid])
-                id = next_id
-            }
-            return `${strings.join('\n')}\n`
-        }
-    }
-}
-
-function MemoryPlot(svg, data, left_pad, colors=schemeTableau10) {
-    function format_points(d) {
-        const size = d.size
-        const xs = d.timesteps.map(t => xscale(t))
-        const bottom = d.offsets.map(t => yscale(t))
-        const m = Array.isArray(size) ? ((t, i) => yscale(t + size[i]))
-                                      :  (t => yscale(t + size))
-        const top = d.offsets.map(m)
-        const p0 = xs.map((x, i) => `${x},${bottom[i]}`)
-        const p1 = xs.map((x, i) => `${x},${top[i]}`).reverse()
-
-        return `${p0.join(' ')} ${p1.join(' ')}`
-    }
-
-    let max_timestep = data.max_at_time.length
-    let max_size = data.max_size
-
-    let width = svg.attr('width')
-    let height = svg.attr('height')
-    let plot_width = width - left_pad
-    let plot_height = height
-
-    let yscale = scaleLinear().domain([0, max_size]).range([plot_height, 0]);
-    let heightscale = scaleLinear().domain([0, max_size]).range([0, plot_height]);
-    let yaxis = axisLeft(yscale).tickFormat(d3.format("~s"))
-    let xscale = scaleLinear().domain([0, max_timestep]).range([0, plot_width])
-    let plot_coordinate_space = svg.append("g").attr("transform", `translate(${left_pad}, ${0})`)
-    let plot_outer = plot_coordinate_space.append('g')
-
-    function view_rect(a) {
-        return a.append('rect').attr('x', 0).attr('y', 0)
-                .attr('width', plot_width).attr('height', plot_height)
-                .attr('fill', 'white')
-    }
-
-    view_rect(plot_outer)
-
-    let cp = svg.append("clipPath").attr("id", "clip")
-    view_rect(cp)
-    plot_outer.attr('clip-path', "url(#clip)")
-
-
-    let zoom_group = plot_outer.append("g")
-    let scrub_group = zoom_group.append('g')
-
-    let plot = scrub_group.selectAll("polygon")
-    .data(data.allocations_over_time)
-    .enter()
-    .append("polygon")
-    .attr('points', format_points)
-    .attr('fill', d => colors[d.color % colors.length])
-
-    let axis = plot_coordinate_space.append('g').call(yaxis)
-
-
-    let scale_mini = 0
-    let translate_mini = 0
-    function handleZoom(e) {
-        const t = e.transform
-        zoom_group.attr("transform", t)
-        axis.call(yaxis.scale(e.transform.rescaleY(yscale)))
-    }
-
-    const thezoom = zoom().on('zoom', handleZoom)
-    plot_outer.call(thezoom)
-
-    return {
-        select_window: (stepbegin, stepend, max) => {
-            let begin = xscale(stepbegin)
-            let size = xscale(stepend) - xscale(stepbegin);
-            let scale = plot_width / size
-            let translate = -begin
-            let yscale =  max_size/max
-            scrub_group.attr("transform", `scale(${scale/yscale}, 1) translate(${translate}, 0)`)
-            plot_outer.call(thezoom.transform, zoomIdentity.scale(yscale).translate(0, -(plot_height - plot_height/yscale)))
-        },
-        set_delegate: (delegate) => {
-            plot.on('mouseover', function (e, d) { delegate.set_selected(d3.select(this)) } )
-            .on('mousedown', function(e, d) { delegate.default_selected = d3.select(this)})
-            .on('mouseleave', function (e, d) { delegate.set_selected(delegate.default_selected) } )
-        }
-    }
-}
-
-function ContextViewer(text, data) {
-    let current_selected = null
-
-    return {
-        default_selected: null,
-        set_selected: (d) => {
-            if (current_selected !== null) {
-                current_selected.attr('stroke', null).attr('stroke-width', null);
-            }
-            if (d === null) {
-                text.text("")
-            } else {
-                const dd = d.datum()
-                if (dd.elem === 'summarized') {
-                    text.html(
-                        "Small tensors that were not plotted to cutdown on render time.\n" +
-                        "Use detail slider to see smaller allocations.")
-                } else {
-                    text.text(`${dd.elem} ${data.context_for_id(dd.elem)}`)
-                }
-                d.attr('stroke', 'black').attr('stroke-width', 1).attr('vector-effect', 'non-scaling-stroke')
-            }
-            current_selected = d
-        }
-    }
-}
-
-
-function MiniMap(mini_svg, plot, data, left_pad, height=70) {
-    let max_at_time = data.max_at_time
-    let width = mini_svg.attr('width')
-    let plot_width = width - left_pad
-    let yscale = scaleLinear().domain([0, data.max_size]).range([height, 0]);
-    let minixscale = scaleLinear().domain([0, max_at_time.length]).range([left_pad, width])
-
-    let mini_points = [[max_at_time.length, 0], [0, 0]]
-
-    for (const [i, m] of max_at_time.entries()) {
-        let [lastx, lasty] = mini_points[mini_points.length - 1]
-        if (m !== lasty) {
-            mini_points.push([i, lasty])
-            mini_points.push([i, m])
-        } else if (i === max_at_time.length - 1) {
-            mini_points.push([i, m])
-        }
-    }
-
-
-    let points = mini_points.map(([t, o]) => `${minixscale(t)}, ${yscale(o)}`)
-    points = points.join(' ')
-    mini_svg.append('polygon').attr('points', points).attr('fill', schemeTableau10[0])
-
-    let xscale = scaleLinear().domain([0, max_at_time.length]).range([0, plot_width])
-
-
-    const brush = brushX()
-    brush.extent([[left_pad, 0], [width, height]])
-    brush.on('brush', function({selection}) {
-        let [begin, end] = selection.map(x => x - left_pad)
-
-        let stepbegin = Math.floor(xscale.invert(begin))
-        let stepend = Math.floor(xscale.invert(end))
-        let max = 0
-        for (let i = stepbegin; i < stepend; i++) {
-            max = Math.max(max, max_at_time[i])
-        }
-        plot.select_window(stepbegin, stepend, max)
-    })
-    mini_svg.call(brush)
-    return {}
-}
-
-function Legend(plot_svg, categories, width) {
-    let xstart = width - 100
-    let ystart = 30
-    plot_svg.append('g').selectAll('rect')
-    .data(categories)
-    .enter()
-    .append('rect')
-    .attr('x', (c, i) => xstart)
-    .attr('y', (c, i) => ystart + i*15)
-    .attr('width', 10)
-    .attr('height', 10)
-    .attr('fill', (c, i) => schemeTableau10[i % schemeTableau10.length])
-    plot_svg.append('g').selectAll('text')
-    .data(categories)
-    .enter()
-    .append('text')
-    .attr('x', (c, i) => xstart + 20)
-    .attr('y', (c, i) => ystart + i*15 + 8)
-    .attr("font-family", "helvetica")
-    .attr('font-size', 10)
-    .text((c) => c)
-    return {}
-}
-
-
-function create(max_entries) {
-    let left_pad = 70
-    let width = 1024
-    let height = 768
-    let data = process_alloc_data(max_entries)
-    let body = d3.select("body")
-    body.selectAll('svg').remove()
-    body.selectAll('div').remove()
-
-    if (alloc_data.elements_info.length > max_entries) {
-         let d = body.append('div')
-         d.append('input')
-         .attr("type", "range")
-         .attr('min', 0)
-         .attr('max', alloc_data.elements_info.length)
-         .attr("value", max_entries)
-         .on('change', function() {
-            create(this.value)
-         })
-         d.append('label').text('Detail')
-    }
-
-    let plot_svg = body.append("svg").attr('width', width).attr('height', height).attr('display', 'block')
-    let plot = MemoryPlot(plot_svg, data, left_pad)
-
-    if (alloc_data.categories !== null) {
-        Legend(plot_svg.append('g'), alloc_data.categories, width)
-    }
-
-    MiniMap(body.append("svg").attr('width', width).attr('height', 80).attr('display', 'block'), plot, data, left_pad)
-    let delegate = ContextViewer(body.append("div").append("pre").text('none'), data)
-    plot.set_delegate(delegate)
-}
-
-create(15000)
-
+main(alloc_data)
 </script>
 </body>
 </html>
@@ -1069,507 +723,9 @@ html, body {
 </head>
 <body>
 <script type="module">
-import * as d3 from "https://cdn.jsdelivr.net/npm/d3@7.7.0/+esm";
-import {schemeTableau10} from "https://cdn.skypack.dev/d3-scale-chromatic@3";
-import {axisLeft} from "https://cdn.skypack.dev/d3-axis@3";
-import {scaleLinear} from "https://cdn.skypack.dev/d3-scale@4";
-import {zoom, zoomIdentity} from "https://cdn.skypack.dev/d3-zoom@3";
-import {brushX} from "https://cdn.skypack.dev/d3-brush@3";
-
+import {main} from "./StatePlot.js"
 let trace_data = $PLOT_DATA
-
-function frames_for(id) {
-    let strings = []
-    while (id !== null) {
-        const [sid, next_id] = trace_data.suffix_table[id]
-        strings.push(trace_data.string_table[sid])
-        id = next_id
-    }
-    return `${strings.join('\n')}\n`
-}
-
-let events_data = trace_data.events
-let string_table = trace_data.string_table
-
-class Event {
-    constructor(id) {
-        this.idx = id
-    }
-    get size() { return events_data.size[this.idx] }
-    get action() { return string_table[events_data.action[this.idx]] }
-    get addr() { return events_data.addr[this.idx] }
-    get frames() { return frames_for(this.frames_idx) }
-    get frames_idx() { return events_data.frames[this.idx]}
-    get stream() { return events_data.stream[this.idx] }
-    get device_free() { return events_data.addr[this.idx] }
-}
-
-function createEvents() {
-    let events = trace_data.events.action.map( (_, i) => new Event(i))
-
-    function version_space() {
-        let version = {}
-        return (addr, increment) => {
-            if (!(addr in version)) {
-                version[addr] = 0
-            }
-            let r = version[addr]
-            if (increment) {
-                version[addr]++
-            }
-            return r
-        }
-    }
-
-    let segment_version = version_space()
-    let block_version = version_space()
-    for (let t of events) {
-        // set unique version for each time an address is used
-        // so that ctrl-f can be used to search for the beginning
-        // and end of allocations and segments
-        switch (t.action) {
-            case 'free':
-            case 'free_completed':
-                t.version = block_version(t.addr, true)
-                break
-            case 'free_requested':
-            case 'alloc':
-                t.version = block_version(t.addr, false)
-                break
-            case 'segment_free':
-            case 'segment_unmap':
-                t.version = segment_version(t.addr, true)
-                break;
-            case 'segment_alloc':
-            case 'segment_map':
-                t.version = segment_version(t.addr, false)
-                break
-            default:
-                break
-        }
-    }
-    trace_data.segment_version = segment_version
-    trace_data.block_version = block_version
-    return events
-}
-
-function Segment(addr, size, stream, frame_idx, version) {
-    return {addr, size, stream, version, get frames() { return frames_for(frame_idx)}}
-}
-function Block(addr, size, real_size, frame_idx, free_requested, version) {
-    console.assert(frame_idx !== undefined)
-    return {addr, size, real_size, get frames() { return frames_for(frame_idx)}, free_requested, version}
-}
-
-function EventSelector(body, outer, events, stack_info, memory_view) {
-    let events_div = outer
-    .append("div")
-    .attr('style', 'grid-column: 1; grid-row: 1; overflow: auto; font-family: monospace')
-
-    let events_selection = events_div
-    .selectAll("pre")
-    .data(events)
-    .enter()
-    .append("pre")
-    .text((e) => formatEvent(e))
-    .attr('style', '')
-
-    let selected_event_idx = null
-
-    let es = {
-        select(idx) {
-            if (selected_event_idx !== null) {
-                let selected_event = d3.select(events_div.node().children[selected_event_idx])
-                selected_event.attr('style', '')
-            }
-            if (idx !== null) {
-                let div = d3.select(events_div.node().children[idx])
-                div.attr('style', `background-color: ${schemeTableau10[5]}`)
-                let [reserved, allocated] = memory_view.draw(idx)
-                let enter = () => eventStack(div.datum(), allocated, reserved)
-                stack_info.highlight(enter)
-                div.node().scrollIntoViewIfNeeded(false)
-            } else {
-                memory_view.draw(0)
-            }
-            selected_event_idx = idx
-        }
-    }
-    body.on('keydown', (e) => {
-        let actions = {ArrowDown: 1, ArrowUp: -1}
-        if (selected_event_idx !== null && e.key in actions) {
-            let new_idx = selected_event_idx + actions[e.key]
-            es.select(Math.max(0, Math.min(new_idx, events.length - 1)))
-            e.preventDefault()
-        }
-    })
-
-    stack_info.register(events_selection, (t) => eventStack(t.datum()), (t) => {}, (d) => es.select(d.datum().idx))
-
-    return es
-}
-
-function formatSize(num) {
-    let orig = num
-    // https://stackoverflow.com/questions/1094841/get-human-readable-version-of-file-size
-    const units = ["", "Ki", "Mi", "Gi", "Ti", "Pi", "Ei", "Zi"];
-    for (const unit of units) {
-        if (Math.abs(num) < 1024.0) {
-        return `${num.toFixed(1)}${unit}B (${orig} bytes)`;
-        }
-        num /= 1024.0;
-    }
-    return `${num.toFixed(1)}YiB`;
-}
-
-function formatEvent(event) {
-    function formatAddr(event) {
-        let version = event.version == 0 ? "" : `_${event.version}`
-        let prefix = event.action.startsWith("segment") ? "s" : "b"
-        return `${prefix}${event.addr.toString(16)}_${event.version}`
-    }
-    let stream = event.stream == 0 ? "" : `\n              (stream ${event.stream})`
-    switch (event.action) {
-        case 'oom':
-            return `OOM (requested ${formatSize(event.size)}, CUDA has ${formatSize(event.device_free)} memory free)${stream}`
-        case 'snapshot':
-            return "snapshot"
-        default:
-            return `${event.action.padEnd(14)} ${formatAddr(event).padEnd(18)} ${formatSize(event.size)}${stream}`
-    }
-}
-
-function eventStack(e, allocated, reserved) {
-    let event = formatEvent(e)
-    if (reserved !== undefined) {
-        event = `(${formatSize(allocated)} allocated / ${formatSize(reserved)} reserved)\n${event}`
-    }
-    return event + "\n" + e.frames
-}
-
-function hashCode(num) {
-  const numStr = num.toString();
-  let hash = 0;
-  for (let i = 0; i < numStr.length; i++) {
-    const charCode = numStr.charCodeAt(i);
-    hash = ((hash << 5) - hash) + charCode;
-    hash = hash & hash; // Convert to 32-bit integer
-  }
-  return hash;
-}
-
-function addStroke(d) {
-    d.attr('stroke', 'red')
-    .attr('stroke-width', '2')
-    .attr('vector-effect', 'non-scaling-stroke')
-}
-
-function removeStroke(d) {
-    d.attr('stroke', '')
-}
-
-function MemoryView(outer, stack_info, trace_data, events) {
-    let svg = outer.append('svg')
-    .attr('style', 'grid-column: 2; grid-row: 1; width: 100%; height: 100%;')
-    .attr('viewBox', '0 0 200 100')
-    .attr('preserveAspectRatio', 'xMinYMin meet')
-    let g = svg.append('g')
-    let seg_zoom = zoom()
-    seg_zoom.on('zoom', ({transform}) => {
-        g.attr('transform', transform)
-    })
-    svg.call(seg_zoom)
-
-    let sorted_segments = []
-    let block_map = {}
-
-    let segments_data = trace_data.segments
-    for (let [i, addr] of trace_data.segments.addr.entries()) {
-        sorted_segments.push(Segment(addr, segments_data.size[i], segments_data.stream[i],
-                                     null, trace_data.segment_version(addr, false)))
-    }
-    sorted_segments.sort((x, y) => x.addr - y.addr)
-
-    let blocks_data = trace_data.blocks
-    for (let [i, addr] of trace_data.blocks.addr.entries()) {
-        block_map[addr] = Block(addr, blocks_data.size[i], blocks_data.real_size[i],
-                                blocks_data.frames[i], blocks_data.pending_free[i],
-                                trace_data.block_version(addr, false))
-    }
-
-    function simulate_memory(idx) {
-        // create a copy of segments because we edit size properties below
-        let l_segments = sorted_segments.map((x) => { return {...x} })
-        let l_block_map = {...block_map}
-
-        function map_segment(merge, seg) {
-            let idx = l_segments.findIndex(e => e.addr > seg.addr)
-            if (!merge) {
-                l_segments.splice(idx, 0, seg)
-                return
-            }
-            if (idx == -1) {
-                idx = l_segments.length
-            }
-            l_segments.splice(idx, 0, seg)
-            if (idx + 1 < l_segments.length) {
-                let next = l_segments[idx + 1]
-                if (seg.addr + seg.size == next.addr && seg.stream == next.stream) {
-                    seg.size += next.size
-                    l_segments.splice(idx + 1, 1)
-                }
-            }
-            if (idx > 0) {
-                let prev = l_segments[idx - 1]
-                if (prev.addr + prev.size == seg.addr && prev.stream == seg.stream) {
-                    prev.size += seg.size
-                    l_segments.splice(idx, 1)
-                }
-            }
-        }
-        function unmap_segment(merge, seg) {
-            if (!merge) {
-                l_segments.splice(l_segments.findIndex(x => x.addr == seg.addr), 1)
-                return
-            }
-            let seg_end = seg.addr + seg.size
-            let idx = l_segments.findIndex(e => e.addr <= seg.addr && seg_end <= e.addr + e.size)
-            let existing = l_segments[idx]
-            let existing_end = existing.addr + existing.size
-            if (existing.addr == seg.addr) {
-                existing.addr += seg.size
-                existing.size -= seg.size
-                if (existing.size == 0) {
-                    l_segments.splice(idx, 1)
-                }
-            } else if (existing_end == seg_end) {
-                existing.size -= seg.size
-            } else {
-                existing.size = seg.addr - existing.addr
-                seg.addr = seg_end
-                seg.size = existing_end - seg_end
-                l_segments.splice(idx + 1, 0, seg)
-            }
-        }
-
-        for (let i = events.length - 1; i > idx; i--) {
-            let event = events[i]
-            switch (event.action) {
-                case 'free':
-                    l_block_map[event.addr] = Block(event.addr, event.size, event.size, event.frames_idx, false, event.version)
-                    break
-                case 'free_requested':
-                    l_block_map[event.addr].free_requested = false
-                    break
-                case 'free_completed':
-                    l_block_map[event.addr] = Block(event.addr, event.size, event.size, event.frames_idx, true, event.version)
-                    break
-                case 'alloc':
-                    delete l_block_map[event.addr]
-                    break
-                case 'segment_free':
-                case 'segment_unmap':
-                    map_segment(event.action == 'segment_unmap',
-                                Segment(event.addr, event.size, event.stream, event.frames_idx, event.version))
-                    break
-                case 'segment_alloc':
-                case 'segment_map':
-                    unmap_segment(event.action == 'segment_map',
-                                  Segment(event.addr, event.size, event.stream, event.frames_idx, event.version))
-                    break
-                case 'oom':
-                    break
-                default:
-                    console.log(`unknown event: ${event.action}`)
-                    break
-            }
-        }
-        let new_blocks = Object.values(l_block_map)
-        return [l_segments, new_blocks]
-    }
-
-    return {
-        draw(idx) {
-            let [segments_unsorted, blocks] = simulate_memory(idx)
-            g.selectAll('g').remove()
-
-            let segment_d = g.append('g')
-            let block_g = g.append('g')
-            let block_r = g.append('g')
-
-            segment_d.selectAll('rect').remove()
-            block_g.selectAll('rect').remove()
-            block_r.selectAll('rect').remove()
-            let segments = [...segments_unsorted].sort((x, y) => x.size == y.size ? (x.addr - y.addr) : (x.size - y.size))
-
-            let segments_by_addr = [...segments].sort((x, y) => x.addr - y.addr)
-
-            let max_size = segments.length == 0 ? 0 : segments.at(-1).size
-
-            let xScale = scaleLinear([0, max_size], [0, 200])
-            let padding = xScale.invert(1)
-
-            let cur_row = 0
-            let cur_row_size = 0
-            for (let seg of segments) {
-                seg.occupied = 0
-                seg.internal_free = 0
-                if (cur_row_size + seg.size > max_size) {
-                    cur_row_size = 0
-                    cur_row += 1
-                }
-                seg.offset = cur_row_size
-                seg.row = cur_row
-                cur_row_size += seg.size + padding
-            }
-
-            let num_rows = cur_row + 1
-
-            let yScale = scaleLinear([0, num_rows], [0, 100])
-
-            let segments_selection = segment_d.selectAll('rect').data(segments).enter()
-            .append('rect')
-            .attr('x', (x) => xScale(x.offset))
-            .attr('y', (x) => yScale(x.row))
-            .attr('width', (x) => xScale(x.size))
-            .attr('height', yScale(4/5))
-            .attr('stroke', 'black')
-            .attr('stroke-width', '1')
-            .attr('vector-effect', 'non-scaling-stroke')
-            .attr('fill', 'white')
-
-            stack_info.register(segments_selection,
-                (d) => {
-                    addStroke(d)
-                    let t = d.datum()
-                    let free = t.size - t.occupied
-                    let internal = ""
-                    if (t.internal_free > 0) {
-                        internal = ` (${t.internal_free/free*100}% internal)`
-                    }
-                    return `s${t.addr.toString(16)}_${t.version}: segment ${formatSize(t.size)} allocated, ` +
-                           `${formatSize(free)} free${internal} (stream ${t.stream})\n${t.frames}`
-                },
-                (d) => {
-                    d.attr('stroke', 'black')
-                    .attr('stroke-width', '1')
-                    .attr('vector-effect', 'non-scaling-stroke')
-                }
-            )
-
-            function find_segment(addr) {
-                let left = 0;
-                let right = segments_by_addr.length - 1;
-                while (left <= right) {
-                    let mid = Math.floor((left + right) / 2);
-                    if (addr < segments_by_addr[mid].addr) {
-                        right = mid - 1;
-                    } else if (addr >= segments_by_addr[mid].addr + segments_by_addr[mid].size) {
-                        left = mid + 1;
-                    } else {
-                        return segments_by_addr[mid];
-                    }
-                }
-                return null;
-            }
-
-            for (let b of blocks) {
-                b.segment = find_segment(b.addr)
-                b.segment.occupied += b.real_size
-                b.segment.internal_free += (b.size - b.real_size)
-            }
-
-            let block_selection = block_g.selectAll('rect').data(blocks).enter()
-            .append('rect')
-            .attr('x', (x) => xScale(x.segment.offset + (x.addr - x.segment.addr)))
-            .attr('y', (x) => yScale(x.segment.row))
-            .attr('width', (x) => xScale(x.real_size))
-            .attr('height', yScale(4/5))
-            .attr('fill', (x, i) => x.free_requested ? 'red' : schemeTableau10[Math.abs(hashCode(x.addr)) % schemeTableau10.length])
-
-            stack_info.register(block_selection, (d) => {
-                addStroke(d)
-                let t = d.datum()
-                let requested = ""
-                if (t.free_requested) {
-                    requested = " (block freed but waiting due to record_stream)"
-                }
-                return `b${t.addr.toString(16)}_${t.version} ` +
-                       `${formatSize(t.real_size)} allocation${requested} (stream ${t.segment.stream})\n` + t.frames
-            }, removeStroke)
-
-            let free_selection = block_r.selectAll('rect').data(blocks).enter()
-            .append('rect')
-            .attr('x', (x) => xScale(x.segment.offset + (x.addr - x.segment.addr) + x.real_size))
-            .attr('y', (x) => yScale(x.segment.row))
-            .attr('width', (x) => xScale(x.size - x.real_size))
-            .attr('height', yScale(4/5))
-            .attr('fill', (x, i) => 'red')
-
-            stack_info.register(free_selection, (d) => {
-                addStroke(d)
-                let t = d.datum()
-                return `Free space lost due to rounding ${formatSize(t.size - t.real_size)}` +
-                       ` (stream ${t.segment.stream})\n` + t.frames
-            }, removeStroke)
-
-            let reserved = segments.reduce((x, y) => x + y.size, 0)
-            let allocated = blocks.reduce((x, y) => x + y.real_size, 0)
-            return [reserved, allocated]
-        }
-    }
-}
-
-function StackInfo(outer) {
-    let stack_trace = outer
-    .append('pre')
-    .attr("style", "grid-column: 1 / 3; grid-row: 2; overflow: auto")
-    let selected = {
-        enter: () => { stack_trace.text("") },
-        leave: () => {},
-    }
-    return {
-        register(dom, enter, leave = (e) => {}, select = (e) => {}) {
-            dom.on('mouseover', (e) => {
-                selected.leave()
-                stack_trace.text(enter(d3.select(e.target)))
-            })
-            .on('mousedown', (e) => {
-                let obj = d3.select(e.target)
-                selected = {enter: () => stack_trace.text(enter(obj)), leave: () => leave(obj)}
-                select(obj)
-            })
-            .on('mouseleave', (e) =>  {
-                leave(d3.select(e.target))
-                selected.enter()
-            })
-        },
-        highlight(enter, leave = () => {}) {
-            selected = {enter: () => stack_trace.text(enter()), leave: leave}
-            selected.enter()
-        }
-    }
-}
-
-function main() {
-    // add some supplement information to trace_data
-    let events = createEvents()
-
-    let body = d3.select("body")
-
-    let outer = body.append("div")
-    .attr('style', "display: grid; grid-template-columns: 1fr 2fr; grid-template-rows: 2fr 1fr; height: 100%; gap: 10px")
-
-    let stack_info = StackInfo(outer)
-    let memory_view = MemoryView(outer, stack_info, trace_data, events)
-    let event_selector = EventSelector(body, outer, events, stack_info, memory_view)
-
-    window.addEventListener('load', function() {
-        event_selector.select(events.length > 0 ? events.length - 1 : null)
-    });
-}
-main()
-
+main(trace_data)
 </script>
 </body>
 </html>
